pyproject.toml
[project]
name = "09-database-sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.116.1",
    "jose>=1.0.0",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
]
